+++ CHANGELOG.md
@@ -10,12 +10,19 @@
   * moved svg background icons to `scss/bg-icons/`
 * updated all dependencies to newest version and improved configs   
 * improved jest code coverage output
+* fixed browserSync page reload when js files are changing
 * added react.js
 * made jsb optional
-* fixed browserSync page reload when js files are changing
+* added foundation css framework
+* reworked included scss mixins
+  * now all chosen mixins gets copied into project
+  * converted extends to mixins
+* added libraries `siema`, `choices`, `react-select`, `a11y-dialog`, `lodash` and `date-fns` 
+* removed libraries `jquery`, `swiper` and `select2`
 * removed `assemble`
 * removed `grunt-accessibility`
 * removed `grunt-htmlhint`
+* removed `grunt-prettify`
 * removed form framework
 * rewrote generator logic completely and improved templates
 * updated generator dependencies